+++ libs/component-sample/src/lib/feature/dialog-sample/dialog-sample.component.ts
@@ -1,13 +1,22 @@
 import { Component, Inject } from '@angular/core';
-import { SdsDialogService, SdsDialogRef, SDS_DIALOG_DATA } from '@gsa-sam/components';
+import {
+  SdsDialogService,
+  SdsDialogRef,
+  SDS_DIALOG_DATA
+} from '@gsa-sam/components';
 
 export interface DialogData {
   animal: string;
   name: string;
 }
 
+export interface AlertData {
+  title: string;
+  content: string;
+}
+
 @Component({
-  selector: 'dialog-overview-example-dialog',
+  selector: 'sds-dialog-sample-data',
   templateUrl: 'dialog-overview-example-dialog.html'
 })
 export class DialogOverviewExampleDialog {
@@ -21,29 +30,61 @@ export class DialogOverviewExampleDialog {
   }
 }
 
+/**
+ * NESTED
+ * ================================================
+ */
 @Component({
-  selector: 'dialog-overview-example-dialog-2',
+  selector: 'sds-dialog-sample-nested',
   template: `
-    <button (click)="openDialog()">Open another dialog</button>
+    <button class="usa-button" (click)="openDialog()">
+      Open another dialog
+    </button>
   `
 })
-export class DialogOverviewExample2Dialog {
+export class NestedDialogComponent {
   animal: string;
   name: string;
   constructor(public dialog: SdsDialogService) {}
   openDialog(): void {
     const dialogRef = this.dialog.open(DialogOverviewExampleDialog, {
-      width: '250px',
+      width: 'small',
       data: { name: this.name, animal: this.animal }
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      console.log('The dialog was closed');
       this.animal = result;
     });
   }
 }
 
+/**
+ * ALERTS
+ * ================================================
+ */
+// Error
+@Component({
+  selector: 'sds-dialog-sample-alert',
+  templateUrl: './alert/template.html'
+})
+export class AlertComponent {
+  constructor(@Inject(SDS_DIALOG_DATA) public data: AlertData) {}
+}
+
+/**
+ * OFFICIAL
+ * ================================================
+ */
+@Component({
+  selector: 'sds-dialog-sample-official',
+  templateUrl: './official/template.html'
+})
+export class OfficialComponent {}
+
+/*
+ * MAIN COMPONENT
+ * ================================================
+ */
 @Component({
   selector: 'sds-modal-sample',
   templateUrl: 'dialog-sample.component.html'
@@ -56,17 +97,30 @@ export class DialogSampleComponent {
 
   openDialog(): void {
     const dialogRef = this.dialog.open(DialogOverviewExampleDialog, {
-      width: '250px',
+      width: 'medium',
       data: { name: this.name, animal: this.animal }
     });
 
     dialogRef.afterClosed().subscribe(result => {
-      console.log('The dialog was closed');
       this.animal = result;
     });
   }
 
-  openDialog2(): void {
-    const dialogRef = this.dialog.open(DialogOverviewExample2Dialog);
+  openNestedDialog() {
+    const dialogRef = this.dialog.open(NestedDialogComponent);
+  }
+
+  openAlert(title, content, alert, size) {
+    const dialogRef = this.dialog.open(AlertComponent, {
+      alert: alert,
+      width: size,
+      data: { title: title, content: content }
+    });
+  }
+
+  openOfficial() {
+    const dialogRef = this.dialog.open(OfficialComponent, {
+      width: 'large'
+    });
   }
 }